Remove conversion warning from src/operation directory 	modified:   src/operation/distance/DistanceOp.cpp 	modified:   src/operation/intersection/RectangleIntersection.cpp 	modified:   src/operation/intersection/RectangleIntersectionBuilder.cpp 	modified:   src/operation/linemerge/LineMerger.cpp 	modified:   src/operation/linemerge/LineSequencer.cpp 	modified:   src/operation/overlay/OverlayOp.cpp 	modified:   src/operation/predicate/RectangleContains.cpp 	modified:   src/operation/valid/IsValidOp.cpp

+++ src/operation/valid/IsValidOp.cpp
@@ -70,7 +70,7 @@ IsValidOp::findPtNotNode(const CoordinateSequence *testCoords,
 	// find a point in the testCoords which is not a node of the searchRing
 	EdgeIntersectionList &eiList=searchEdge->getEdgeIntersectionList();
 	// somewhat inefficient - is there a better way? (Use a node map, for instance?)
-	unsigned int npts=testCoords->getSize();
+	unsigned int npts = static_cast<int>(testCoords->getSize());
 	for(unsigned int i=0; i<npts; ++i)
 	{
 		const Coordinate& pt=testCoords->getAt(i);
@@ -231,7 +231,7 @@ IsValidOp::checkValid(const Polygon *g)
 void
 IsValidOp::checkValid(const MultiPolygon *g)
 {
-	unsigned int ngeoms = g->getNumGeometries();
+	unsigned int ngeoms = static_cast<unsigned int>(g->getNumGeometries());
 	vector<const Polygon *>polys(ngeoms);
 
 	for (unsigned int i=0; i<ngeoms; ++i)
@@ -284,7 +284,7 @@ IsValidOp::checkValid(const MultiPolygon *g)
 void
 IsValidOp::checkValid(const GeometryCollection *gc)
 {
-	for(unsigned int i=0, ngeoms=gc->getNumGeometries(); i<ngeoms; ++i)
+	for(unsigned int i=0, ngeoms=static_cast<unsigned int>(gc->getNumGeometries()); i<ngeoms; ++i)
 	{
 		const Geometry *g=gc->getGeometryN(i);
 		checkValid(g);
@@ -382,7 +382,7 @@ IsValidOp::checkHolesInShell(const Polygon *p, GeometryGraph *graph)
 	const LinearRing *shell=static_cast<const LinearRing*>(
 			p->getExteriorRing());
 
-	int nholes = p->getNumInteriorRing();
+	int nholes = static_cast<int>(p->getNumInteriorRing());
 
 	if(shell->isEmpty())
 	{
@@ -446,7 +446,7 @@ IsValidOp::checkHolesNotNested(const Polygon *p, GeometryGraph *graph)
 	//QuadtreeNestedRingTester nestedTester(graph);
 	IndexedNestedRingTester nestedTester(graph);
 
-	int nholes=p->getNumInteriorRing();
+	int nholes = static_cast<int>(p->getNumInteriorRing());
 	for(int i=0; i<nholes; ++i)
 	{
 		assert(dynamic_cast<const LinearRing*>(
@@ -474,7 +474,7 @@ IsValidOp::checkHolesNotNested(const Polygon *p, GeometryGraph *graph)
 void
 IsValidOp::checkShellsNotNested(const MultiPolygon *mp, GeometryGraph *graph)
 {
-	for(unsigned int i=0, ngeoms = mp->getNumGeometries(); i<ngeoms; ++i)
+	for(unsigned int i=0, ngeoms = static_cast<int>(mp->getNumGeometries()); i<ngeoms; ++i)
 	{
 		const Polygon *p=dynamic_cast<const Polygon *>(
 				mp->getGeometryN(i));
@@ -524,7 +524,7 @@ IsValidOp::checkShellNotNested(const LinearRing *shell, const Polygon *p,
 	if (!insidePolyShell) return;
 
 	// if no holes, this is an error!
-	int nholes = p->getNumInteriorRing();
+	int nholes = static_cast<int>(p->getNumInteriorRing());
 	if (nholes<=0) {
 		validErr=new TopologyValidationError(
 			TopologyValidationError::eNestedShells,
@@ -604,7 +604,7 @@ IsValidOp::checkConnectedInteriors(GeometryGraph &graph)
 void
 IsValidOp::checkInvalidCoordinates(const CoordinateSequence *cs)
 {
-	unsigned int size=cs->getSize();
+	unsigned int size = static_cast<unsigned int>(cs->getSize());
 	for (unsigned int i=0; i<size; ++i)
 	{
 		if (! isValid(cs->getAt(i)) )
@@ -625,7 +625,7 @@ IsValidOp::checkInvalidCoordinates(const Polygon *poly)
 	checkInvalidCoordinates(poly->getExteriorRing()->getCoordinatesRO());
 	if (validErr != nullptr) return;
 
-	int nholes=poly->getNumInteriorRing();
+	int nholes = static_cast<int>(poly->getNumInteriorRing());
 	for (int i=0; i<nholes; ++i)
 	{
 		checkInvalidCoordinates(
@@ -643,7 +643,7 @@ IsValidOp::checkClosedRings(const Polygon *poly)
 	checkClosedRing(lr);
 	if (validErr) return;
 
-	int nholes=poly->getNumInteriorRing();
+	int nholes = static_cast<int>(poly->getNumInteriorRing());
 	for (int i=0; i<nholes; ++i)
 	{
 		lr=(const LinearRing *)poly->getInteriorRingN(i);
